# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OSM2CAI",
        "description": "Api documentation",
        "contact": {
            "email": "catastorei@cai.it"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "http://localhost:8000/assets/images/logo_sosec.jpg"
        }
    },
    "paths": {
        "/api/v1/hiking-routes/region/{region_code}/{sda}": {
            "get": {
                "tags": [
                    "hiking-routes"
                ],
                "operationId": "d6b251f086821ba0af123f1f21b6cb7e",
                "parameters": [
                    {
                        "name": "region_code",
                        "in": "path",
                        "description": "\nRegione code according to CAI convention: <br/>\n<br />A -> Friuli Venezia Giulia\n<br />B -> Veneto\n<br />C -> Trentino Alto Adige\n<br />D -> Lombardia\n<br />E -> Piemonte\n<br />F -> Val d'Aosta\n<br />G -> Liguria\n<br />H -> Emilia Romagna\n<br />L -> Toscana\n<br />M -> Marche\n<br />N -> Umbria\n<br />O -> Lazio\n<br />P -> Abruzzo\n<br />Q -> Molise\n<br />S -> Campania\n<br />R -> Puglia\n<br />T -> Basilicata\n<br />U -> Calabria\n<br />V -> Sicilia\n<br />Z -> Sardegna",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "varchar"
                        }
                    },
                    {
                        "name": "sda",
                        "in": "path",
                        "description": "SDA (stato di accatastamento) (e.g. 3 or 3,1 or 0,1,2). SDA=3 means ready to be validated, SDA=4 means validated by CAI expert",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "varchar"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns all the hiking routes OSM2CAI IDs based on the given region code and SDA number. \n     *                       These ids can be used in the geojson API hiking-route",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "Internal osm2cai Identifier",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object",
                                    "example": [
                                        1269,
                                        652,
                                        273
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/hiking-routes-osm/region/{region_code}/{sda}": {
            "get": {
                "tags": [
                    "hiking-routes-osm"
                ],
                "operationId": "bc33a73211d35e85ee642851e575cd23",
                "parameters": [
                    {
                        "name": "region_code",
                        "in": "path",
                        "description": "\nRegione code according to CAI convention: <br/>\n<br />A -> Friuli Venezia Giulia\n<br />B -> Veneto\n<br />C -> Trentino Alto Adige\n<br />D -> Lombardia\n<br />E -> Piemonte\n<br />F -> Val d'Aosta\n<br />G -> Liguria\n<br />H -> Emilia Romagna\n<br />L -> Toscana\n<br />M -> Marche\n<br />N -> Umbria\n<br />O -> Lazio\n<br />P -> Abruzzo\n<br />Q -> Molise\n<br />S -> Campania\n<br />R -> Puglia\n<br />T -> Basilicata\n<br />U -> Calabria\n<br />V -> Sicilia\n<br />Z -> Sardegna",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "varchar"
                        }
                    },
                    {
                        "name": "sda",
                        "in": "path",
                        "description": "SDA (stato di accatastamento) (e.g. 3 or 3,1 or 0,1,2). SDA=3 means ready to be validated, SDA=4 means validated by CAI expert",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "varchar"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns all the hiking routes OSM IDs based on the given region code and SDA number.\n     *                       OSMID can be used in hiking-route-osm API or directly in OpenStreetMap relation by the following URL:\n     *                       https://openstreetmap.org/relation/{OSMID}. Remember that the data on OSM can be differente from data in \n     *                       OSM2CAI after validation.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "OSM": {
                                            "description": "Open Streen Map identification",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object",
                                    "example": [
                                        7766787,
                                        3151885,
                                        2736729
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/hiking-routes/bb/{bounding_box}/{sda}": {
            "get": {
                "tags": [
                    "hiking-routes-bb"
                ],
                "operationId": "9f5e516865448220fcf8ef960334f3cb",
                "parameters": [
                    {
                        "name": "bounding_box",
                        "in": "path",
                        "description": "List of WGS84 lat,lon cordinates in this order(xmin,ymin,xmax,ymax)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "varchar"
                        }
                    },
                    {
                        "name": "sda",
                        "in": "path",
                        "description": "SDA (stato di accatastamento) (e.g. 3 or 3,1 or 0,1,2). SDA=3 means ready to be validated, SDA=4 means validated by CAI expert",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "varchar"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns all the hiking routes OSM2CAI IDs based on the given bounding box coordinates( xmin,ymin,xmax,ymax)  and SDA number.\n     *                       These ids can be used in the geojson API hiking-route",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "Internal osm2cai Identifier",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object",
                                    "example": [
                                        1269,
                                        652,
                                        273
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}